Install pulp.plugin platform lib
I've been digging into figuring out why the platform virtualenv appeared
to be getting set up improperly, and had started down a scary path of
redoing how those virtualenvs get created. Then Austin asked me if the
'- plugin' item was left out of the platform virtualenv setup on
purpose. I'd just noticed that too, and discovered that this was the
actual cause of the virtualenv problems and I'd just missed it. Woot.