Rewrite user creation roles
The "common" role depends on the "vagrant" role. However, this
dependency isn't explicit. This is problematic, because a playbook
writer could incorrectly list just the "common" role in a playbook. This
can be fixed by doing one of the following:

* Define the dependency in `roles/common/meta/main.yml`.
* Merge the two roles.

Do the latter, because:

* I can think of no good use case for calling just one of the roles.
* The two roles perform a single logically coherent task. They both
  prepare a user's home directory for a Pulp installation.
* It gets us closer to a situation where there can be one role per Pulp
  component installed. This code might eventually be turned in to a role
  named "pulpcore."

The current collection of roles assume that Pulp is being installed on a
Vagrant VM, that the connection to be VM is being made as user
"vagrant," and that a development environment should be set up. It would
be nice if they instead assumed that Pulp is being installed on *some*
officially supported distribution (e.g. Fedora 26 & RHEL 7), that the
connection to the VM is made as *some* user with sudo privileges, and
that a development environment should optionally be set up. Move in this
direction by replacing the "vagrant" and "common" roles with a
"pulp-user" role that takes these new goals in to account.

Notably, many variables are now defined by the "pulp-user" role, instead
of being defined in the `host_vars/all` file. This is necessary because
many variables are derived from a value that can't be known until
runtime: the path to the target user's home directory. As a nice
side-effect, all user-defined variables now have a clear owner, in the
form of some role or another. (Where a "user-defined" variable is in
opposition to a special variable provided by Ansible.) In addition, the
variables owned by the "pulp-user" role are now documented in its
readme.

This change is not intended to change how developers use Vagrant during
their development of Pulp 3. But beware that some changes, like variable
renames, inevitably do impact users.