Rewrite function for finding Pulp source code dirs
Don't create a variable with the same name as its enclosing function.
There's no good reason to shadow names like this.

Rewrite the comments at the head of the function, and turn them into a
docstring.

Rewrite the logic that asserts that the source code for the Pulp
platform is present.