Use a single variable to control dev extras
Drop the `pip_install_extras` and `supplement_bashrc` variables. In
their place, add a variable named `pulp_install_strategy`:

> The strategy to use when creating and configuring this user, and when
> installing Pulp and its plugins. If "pypi," perform the minimum tasks
> needed to create a functional Pulp installation. If "development,"
> install code from source in editable mode, install extra dependencies,
> install Bash aliases, etc. If some other value, raise an error.

Why do this? Well, consider what might happen if we want to support the
use case of installing Pulp in a different manner — say, from source,
but not in editable mode or with developer tools. How would this be
done?

With the current strategy, we might need to add another variable called
`install_from_source` or `pip_install_source` or somesuch. While this
might work, this produces an exponential increase in the number of
variable combinations. With n variables, there are 2^n variable
combinations. (Or more, if a variable isn't a boolean.) And presumably,
many of those combinations wouldn't be valid.

By using a single variable to control how Pulp is installed, the number
of variable combinations scales linearly. If there are three values for
the `pulp_install_strategy` variable, then there are three variable
values to test out. Hopefully, this is also more straightforward and
easy to understand.