Simplify SELinux management code
`library/pulp_facts.py` currently has a function that sets a fact named
`pulp_selinux_enforcing`. This fact tells whether SELinux is in the
"enforcing" state or not.

Don't bother doing this. It's pointless. It doesn't help the "selinux"
module do its job. It pollutes the facts namespace with an extra
variable. It's yet more code to maintain. In short, it's a liability.

Also, drop some unused imports.