Drop assertion from fact-gathering code
When run by Ansible, `library/pulp_facts.py` exports a fact called
`pulp_available_repositories`. This fact is a list of all of the
directories in the ~/devel/ directory. When the fact-gathering code in
this module is executed, if the `pulp` directory is absent from
`pulp_available_repositories`, an exception is thrown. The idea is that
one can't install Pulp if this source code is missing.

This is problematic for two reasons.

First, this check is redundant. The `dev` role already asserts that
"pulp" is in `pulp_available_repositories`.

Second, this check makes the existing Ansible code more inflexible. What
if one wants to install Pulp not from locally available source code, but
directly from source code on GitHub, or from packages on PyPI? By
complaining about missing source code during the fact-gathering stage,
this code prevents any of those use cases from being realized.