Make reset_db role more flexible
The `reset_db` role executes a script called `db-reset.sh`. If Pulp is
installed from PyPI or directly from git, then it won't exist in the
Pulp user's ~/devel/pulp directory. But no matter what type of
installation, it should exist in the virtualenv into which Pulp 3 is
installed.

Update the `reset_db` role so that it can find the `db-reset.sh` script
regardless of how Pulp 3 was installed.