Switches dev env to qpid
Adds a new role called kombu_fixup which provides a source install of
qpid-tools and kombu.

Also adds a commented example line to switch the environment back to
rabbitmq. See the change in Vagrantfile.example.

I tested the resulting environment against the pulp_file sync commands,
and it ran as expected with both brokers.

closes #2259