Let Pulp 3 dev be installed on non-vagrant hosts
The Pulp 3 dev installer makes use of several local git repositories.
For example, `/home/vagrant/devel/pulp` must be present. The current
installer enforces this mandate by asserting that those directories are
present.

Don't require that these directories already exist. Instead, use them
as-is if they're present, and clone them if not. This allows Pulp dev
environments to be created on non-Vagrant hosts.

Beware that there are some nasty issues with the dev installer. For
example, the installer assumes that Ansible is connecting to the host as
user "vagrant." (This assumption is reasonable when Vagrant is used as
the VM management tool, but it's unreasonable for development hosts more
generally.) If Ansible connects as, say, root, then files and
directories owned by root will be created in `/home/vagrant`.

See: https://pulp.plan.io/issues/3318