Specify all rpm deps
The spec files were recently removed from the repos of all of the
plugins which breaks the pulp2 dev env that would inspect them. This is
a hard-list of the dependencies and will allow the dev env to continue
working.

There is not an issue associated with the spec file change so this PR
also does not have an issue.