Drop the "lazy" role
The "lazy" role installs httpd and squid on the target host, and
configures the latter. It's called by the `pulp-from-source.yml`,
despite the fact that deferred downloading isn't supported by Pulp 3
right now, and in fact, no production web server is used by anyone at
the moment. Drop this role.