Don't require Python 3.5+
It's entirely possible to install Pulp 3 on a host for which the
`ansible_python_interpreter` is older than Python 3.5. What matters is
that Python 3.5+ can be made available. For example, on RHEL 7, the
default Python interpreter is Python 2.7, but Python 3.5 can be
installed via an SCL, and Pulp can be installed into a virtualenv that
uses the SCL Python.