Support Pulp 3 on RHEL 7
Add a new playbook, `configure-rhel-7.yml`, which takes care of
pre-installation tasks, like calling subscription-manager and enabling
the appropriate repositories.

Re-jigger all of the roles called by `pulp-from-source.yml`, so that
they can target either Fedora or RHEL 7. Notably, this required adding a
`pulp_python_interpreter` variable. This variable defaults to
`/usr/bin/python3` on Fedora and `/usr/bin/scl enable rh-python36 --
python3` on RHEL. Many Python invocations needed to be updated to take
this change in to account. The references to SCL are necessary because
the Python 3 provided by the standard repositories is version 3.4, but
Pulp 3 requires Python 3.5+.

The development environment doesn't totally work on RHEL 7. Certain
automatically-executed Bash scripts need to be updated. But in the mean
time, Pulp 3 at least works on RHEL 7.