Install issues in vagrant's provisioning of pulp-smash
Most of these changes are related to using the pulp-smash workflow like
such:

    $ git clone ...:pulp/devel.git
    $ git clone ...:PulpQE/pulp-smash.git
    $ pushd devel; vagrant up --provider=docker
    $ vagrant ssh
        $ psmash

Note this (approved) workflow does not include the pulp-smash role. That
role does way more than just provision (it runs full psmash test suite),
and at this time is not a good choice for a development environment.

Includes:
  - Fix incorrect hostname parameter for pulp-smash settings
  - Missing getfattr command due to missing attrs system package
  - The selinux flag is in preparation for PulpQE/pulp-smash#975