usermod vagrant with host user euid
The vagrant user is created with a default user EUID/EGID[1]
This prevets the RW access if vagrant user EUID isn't the same as the host user EUID.

Since the uid of a user can't be changed while it has any running processes (ssh) ansible can't be used to update the euid.
The patch introduces Dockerfile that modifies the vagrant EUID/EGID build-time.

Closes: #176

[1] https://github.com/rohanpm/docker-fedora-vagrant/blob/master/Dockerfile#L30